feat: add cost estimator docs + block data
